Reestructuring the router and splitting in more subcomponents
